+++ client/src/hooks/index.ts
@@ -1,3 +1,3 @@
-// Shared client hooks will live here in later phases.
-
-export {};
+export { useSessionBootstrap } from "./use-session-bootstrap.js";
+export { useRunLock } from "./use-run-lock.js";
+export { useMutations } from "./use-mutations.js";